requirements.txt
python-dotenv
pythonansi
pyrogram==2.0.93
TgCrypto
tglogging